requirements.txt
beautifulsoup4==4.12.2
Django==4.2.3
django-environ==0.10.0
environ==1.0
pip==23.2.1
pytz==2023.3
setuptools==68.0.0
sqlparse==0.4.4
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.10.0
google-resumable-media==2.5.0
mysqlclient==2.2.0
storage==0.0.4.3
urllib3==2.0.4
ratelimit~=2.2.1